fix: 修复装备词缀格式化问题并添加背包同步功能
修正装备词缀的字符串格式化问题，更新配置文件中的百分比显示格式为正确的转义写法，统一词缀类的格式化参数类型确保显示正常。
新增背包界面与玩家角色装备的同步逻辑，实现装备槽与布局ID的转换、加载角色已装备物品以及同步各类装备操作的功能。

+++ app/src/main/java/com/example/treasure_and_battle/affix/impl/equip/trigger/EquipTriggerRecoverAffix.java
@@ -97,7 +97,7 @@ public class EquipTriggerRecoverAffix extends BaseEquipAffix {
         if (damageToRecoverRatio > 0f) {
             return String.format(description, affixValue * 100f, damageToRecoverRatio * 100f);
         }
-        return String.format(description, affixValue * 100f, recoverValue);
+        return String.format(description, affixValue * 100f, (float) recoverValue);
     }
 }
